Backend/Academic_Schedular/requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyPDF2==1.26.0
pdfplumber==0.5.28
pytesseract==0.3.8
python-dotenv==0.19.0